Rework emsdk_env.sh script for bash, zsh and ksh
This version reworks the script to for 3 current
working shells and makes it easier to add more if they
can be made to work.

This also allows the script to work if the emsdk
directory or the emsdk_env.sh script itself are sym links.

this also adds a new test in scripts directory that
tests the 3 know shells, the 3 choices of path to the
script (direct, symlinked dir, symlinked file) and using
both the "source" and "." syntax.